Split the grind shops into folders, and cap what a command file may weigh
/fish, /hunt and /mine became folders in #721, but the split stopped at the
verb group and the groups kept growing. The three shops were the largest
files left in the repo — fish 822 lines, hunt 828, mine 635 — and a shop
that sells rods or weapons or pickaxes, plus upgrades, plus ammo, plus
consumables, plus repairs, plus locations is six unrelated flows sharing
nothing but the guild lookup at the top of the dispatch. Each shop is now
its own folder: an index.js that resolves the guild, the user and the
currency, and one file per verb. Every handler body is unchanged, byte for
byte; mine's seven  blocks became seven functions taking
the user and currency the dispatch already had.

The loader is fine with a folder inside a folder — only
<category>/<name>/index.js registers — so nothing here becomes a command
of its own.

The staged loot reveal existed three times, once per grind: the same tier
thresholds, the same three 1500 ms beats, the same embed shapes, differing
only in flavour text, and untested in all three copies because each sat
behind a real cast in a 700-line command. It is one module with the copy
as a table, which is also the only way to see the three sets of lines
beside each other.

Then the ratchet, because none of this holds by itself.
eslint-rules/command-file-size.js caps a command file at 900 lines — the
number utils/embedColors already jokes about — and reports on the line
that broke it. The four commands that were already over it are frozen at
the length they had today: they may shrink, never grow, and an entry that
reaches the cap is itself an error, because a stale exemption is
permission to grow back into it. tests/grindCommandLayout holds the three
split folders to a tighter 750, down from 900, to bank what the split
bought.

One hazard the deeper nesting introduced is worth naming: every guard that
reads a grind command's source goes through tests/helpers/grindSources,
whose walk was a single readdir. Left alone it would have dropped the
whole shop — 800-odd lines of charges and transactions — out of every one
of those guards, silently, while staying green. It recurses now, and
tests/grindCommandLayout pins that it does.

Closes #917
