+++ src/commands/economy/fish/shop/list.js
@@ -1,0 +1,96 @@
+'use strict';
+
+// `/fish shop list` — the five browse pages.
+
+const {
+    ROD_TIERS,
+    ROD_UPGRADES,
+    BAIT_PACKS,
+    CONSUMABLES,
+    LOCATION_LIST,
+} = require('../../../../data/fishData');
+const { runShopBrowse } = require('../../../../utils/shopBrowse');
+
+async function showShopList(interaction, user, currency) {
+    const f = user.fishing;
+
+    const rodItems = ROD_TIERS.map(r => ({
+        imageId: `fish:${r.slug}`,
+        name:    r.name,
+        price:   r.cost,
+        emoji:   r.emoji,
+        badge:   `T${r.tier}`
+    }));
+    const rodList = ROD_TIERS.map(r =>
+        `${r.emoji} **T${r.tier} ${r.name}** — ${currency}${r.cost.toLocaleString()} · \`/fish shop rod type:${r.slug}\``
+    ).join('\n');
+
+    const upgradeItems = Object.values(ROD_UPGRADES).map(u => ({
+        imageId: `fish:${u.id}`,
+        name:    u.name,
+        emoji:   u.emoji,
+        subline: `~${Math.round(u.costMultiplier * 100)}% of rod`
+    }));
+    const upgradeList = Object.values(ROD_UPGRADES).map(u =>
+        `${u.emoji} **${u.name}** — *${u.description}* · \`/fish shop upgrade type:${u.id}\``
+    ).join('\n');
+
+    const baitItems = BAIT_PACKS.map(p => ({
+        imageId: `fish:${p.id}`,
+        name:    p.name,
+        price:   p.cost,
+        emoji:   p.emoji
+    }));
+    const baitList = BAIT_PACKS.map(p =>
+        `${p.emoji} **${p.name}** — ${currency}${p.cost} · \`/fish shop buy item:${p.id}\``
+    ).join('\n');
+
+    const consumableItems = Object.values(CONSUMABLES).map(c => ({
+        imageId: `fish:${c.id}`,
+        name:    c.name,
+        price:   c.cost,
+        emoji:   c.emoji
+    }));
+    const consumableList = Object.values(CONSUMABLES).map(c =>
+        `${c.emoji} **${c.name}** — ${currency}${c.cost} · \`/fish shop buy item:${c.id}\``
+    ).join('\n');
+
+    const locationItems = LOCATION_LIST.map(loc => {
+        const unlocked = f.unlockedLocations.includes(loc.id);
+        const isActive = f.activeLocation === loc.id;
+        return {
+            imageId: `fish:${loc.id}`,
+            name:    loc.name,
+            emoji:   loc.emoji,
+            badge:   isActive ? 'ACTIVE' : (unlocked ? 'OWNED' : `Lv.${loc.unlockLevel}`),
+            subline: unlocked ? (isActive ? 'Currently fishing' : 'Unlocked') : (loc.unlockCost > 0 ? `${currency}${loc.unlockCost.toLocaleString()}` : 'Free')
+        };
+    });
+    const locationList = LOCATION_LIST.map(loc => {
+        const unlocked = f.unlockedLocations.includes(loc.id);
+        const isActive = f.activeLocation === loc.id;
+        const status = unlocked
+            ? (isActive ? '✅ **ACTIVE**' : '✅ Unlocked')
+            : `🔒 Lv.${loc.unlockLevel}${loc.unlockCost > 0 ? ` / ${currency}${loc.unlockCost.toLocaleString()}` : ' (free)'}`;
+        return `${loc.emoji} **${loc.name}** — ${status}`;
+    }).join('\n');
+
+    return runShopBrowse(interaction, {
+        activity: 'fish',
+        title:    'Fishing Shop',
+        currency,
+        // Activity images are per guild since #561; without this the browse view
+        // only ever finds the shared pre-#561 rows.
+        guildId:  interaction.guild.id,
+        footer:   'rod • upgrade • buy • use • repair • unlock',
+        pages: [
+            { id: 'rods',        label: 'Rods',        emoji: '🎣',  subtitle: 'Better rods, better catches.',           items: rodItems,        listText: rodList        },
+            { id: 'upgrades',    label: 'Upgrades',    emoji: '🔧',  subtitle: 'One module per rod, permanent.',         items: upgradeItems,    listText: upgradeList    },
+            { id: 'bait',        label: 'Bait',        emoji: '🪱',  subtitle: 'The right bait pulls the right fish.',   items: baitItems,       listText: baitList       },
+            { id: 'consumables', label: 'Consumables', emoji: '🧪',  subtitle: 'Luck, XP and quick boosts.',             items: consumableItems, listText: consumableList },
+            { id: 'locations',   label: 'Locations',   emoji: '🗺️', subtitle: 'New waters, new species.',                items: locationItems,   listText: locationList   }
+        ]
+    });
+}
+
+module.exports = { showShopList };
