+++ src/commands/economy/fish/shop/upgrade.js
@@ -1,0 +1,143 @@
+'use strict';
+
+// `/fish shop upgrade` — installing a module on the equipped rod.
+
+const {
+    MessageFlags,
+    EmbedBuilder,
+    ActionRowBuilder,
+    ButtonBuilder,
+    ButtonStyle,
+} = require('discord.js');
+const User = require('../../../../models/User');
+const { attachGrind } = require('../../../../utils/grindProfile');
+const { ensureFishingData } = require('../../../../services/fishService');
+const { ROD_UPGRADES, ROD_BY_TIER } = require('../../../../data/fishData');
+const COLORS = require('../../../../utils/embedColors');
+
+async function handleBuyUpgrade(interaction, user, currency) {
+    const f = user.fishing;
+
+    if (f.equippedRodIndex < 0 || !f.rods[f.equippedRodIndex]) {
+        return interaction.reply({ content: `You don't have a rod equipped. Use \`/fish inv equip\` first.`, flags: MessageFlags.Ephemeral });
+    }
+
+    const upgradeId  = interaction.options.getString('type');
+    const upgradeDef = ROD_UPGRADES[upgradeId];
+
+    if (!upgradeDef) {
+        return interaction.reply({ content: 'Unknown upgrade.', flags: MessageFlags.Ephemeral });
+    }
+
+    const targetRodIndex = f.equippedRodIndex;
+    const rod     = f.rods[targetRodIndex];
+    const rodData = ROD_BY_TIER[rod.tier];
+    const cost    = Math.round(rodData.cost * upgradeDef.costMultiplier);
+
+    if (rod.upgrade) {
+        return interaction.reply({
+            content: `Your **${rod.name}** already has the **${rod.upgrade.replace(/_/g, ' ')}** upgrade. Each rod can only hold one upgrade.`,
+            flags: MessageFlags.Ephemeral
+        });
+    }
+    if (user.balance < cost) {
+        return interaction.reply({
+            content: `You need **${currency}${cost.toLocaleString()}** to install **${upgradeDef.name}**. You have **${currency}${user.balance.toLocaleString()}**.`,
+            flags: MessageFlags.Ephemeral
+        });
+    }
+
+    const confirmEmbed = new EmbedBuilder()
+        .setColor(COLORS.RARE)
+        .setTitle(`${upgradeDef.emoji} Install ${upgradeDef.name}?`)
+        .setDescription(`Installing on **${rod.name}**\n${upgradeDef.description}`)
+        .addFields(
+            { name: 'Cost',         value: `${currency}${cost.toLocaleString()}`,      inline: true },
+            { name: 'Your Balance', value: `${currency}${user.balance.toLocaleString()}`, inline: true }
+        )
+        .setFooter({ text: 'One upgrade per rod. This cannot be removed.' });
+
+    const row = new ActionRowBuilder().addComponents(
+        new ButtonBuilder().setCustomId('upgrade_confirm').setLabel('Install').setStyle(ButtonStyle.Success).setEmoji('✅'),
+        new ButtonBuilder().setCustomId('upgrade_cancel').setLabel('Cancel').setStyle(ButtonStyle.Secondary).setEmoji('❌')
+    );
+
+    const reply = await interaction.reply({ embeds: [confirmEmbed], components: [row], flags: MessageFlags.Ephemeral, fetchReply: true });
+    const collector = reply.createMessageComponentCollector({ time: 30_000 });
+
+    collector.on('collect', async btn => {
+        if (btn.user.id !== interaction.user.id) {
+            return btn.reply({ content: 'This is not your confirmation.', flags: MessageFlags.Ephemeral });
+        }
+        collector.stop();
+
+        if (btn.customId === 'upgrade_cancel') {
+            return btn.update({ content: 'Installation cancelled.', embeds: [], components: [] });
+        }
+
+        const freshUser = await User.findOne({ userId: interaction.user.id, guildId: interaction.guild.id });
+        await attachGrind(freshUser);
+        ensureFishingData(freshUser);
+
+        const freshRod = freshUser.fishing.rods[targetRodIndex];
+        if (!freshRod) {
+            return btn.update({ content: 'That rod is no longer in your inventory.', embeds: [], components: [] });
+        }
+        if (freshRod.upgrade) {
+            return btn.update({ content: `**${freshRod.name}** already has an upgrade installed.`, embeds: [], components: [] });
+        }
+
+        // Guarded atomic debit for the same reason as the rod purchase above: this
+        // lands up to 30 seconds after the prompt, and a `save()` would write an
+        // absolute balance over anything spent in between. Charged only once the
+        // rod is known to be upgradeable, so a rejected install costs nothing.
+        const debited = await User.findOneAndUpdate(
+            { userId: interaction.user.id, guildId: interaction.guild.id, balance: { $gte: cost } },
+            { $inc: { balance: -cost } },
+            { new: true, projection: { balance: 1 } },
+        );
+        if (!debited) {
+            return btn.update({ content: 'Insufficient funds.', embeds: [], components: [] });
+        }
+
+        freshUser.balance = debited.balance;
+        freshUser.unmarkModified('balance');
+        freshRod.upgrade   = upgradeId;
+        freshUser.markModified('fishing');
+
+        try {
+            await freshUser.save();
+        } catch (err) {
+            console.error('[fishshop upgrade] save error:', err);
+            // The coins are already gone; hand them back rather than charging for
+            // an upgrade that was never installed.
+            await User.updateOne(
+                { userId: interaction.user.id, guildId: interaction.guild.id },
+                { $inc: { balance: cost } },
+            ).catch(refundErr => console.error('[fishshop upgrade] refund after failed save:', refundErr));
+            return btn.update({ content: 'Something went wrong and your coins were refunded. Please try again.', embeds: [], components: [] });
+        }
+
+        return btn.update({
+            embeds: [
+                new EmbedBuilder()
+                    .setColor(COLORS.SUCCESS)
+                    .setTitle(`${upgradeDef.emoji} ${upgradeDef.name} Installed!`)
+                    .setDescription(`**${freshRod.name}** now has **${upgradeDef.name}** installed permanently.`)
+                    .addFields(
+                        { name: 'Effect',  value: upgradeDef.description,                       inline: true },
+                        { name: 'Balance', value: `${currency}${freshUser.balance.toLocaleString()}`, inline: true }
+                    )
+            ],
+            components: []
+        });
+    });
+
+    collector.on('end', (_, reason) => {
+        if (reason === 'time') {
+            interaction.editReply({ content: 'Installation timed out.', embeds: [], components: [] }).catch(() => {});
+        }
+    });
+}
+
+module.exports = { handleBuyUpgrade };
