+++ src/commands/economy/hunt/start.js
@@ -42,52 +42,7 @@ const { saveWithBalanceDelta } = require('../../../utils/balanceDelta');
 const { pickApproachProfile, runAimPhase } = require('./aim');
 const { buildBonusLines, buildHuntEmbed } = require('./embeds');
 const { ownedBy } = require('../../../utils/collectorOwner');
-
-// ─── Staged loot reveal for rare+ drops ──────────────────────────────────────
-// tier: 'common'|'uncommon'|'rare'|'epic'|'legendary'|'event'|null
-async function stagedLootReveal(interaction, tier, finalEmbed) {
-    const tierNum = TIER_NUM[tier] ?? 0;
-    if (tierNum < 3) {
-        await interaction.editReply({ embeds: [finalEmbed] });
-        return;
-    }
-    const wait = ms => new Promise(r => setTimeout(r, ms));
-
-    const fogEmbed = new EmbedBuilder()
-        .setColor('#4a4a4a')
-        .setTitle('🌫️ Something stirs in the shadows...')
-        .setDescription('━━━━━━━━━━━━━━━\n*The shadows shift. Something is here.*\n━━━━━━━━━━━━━━━');
-
-    if (tierNum === 3) {
-        await interaction.editReply({ embeds: [fogEmbed] });
-        await wait(1500);
-    } else {
-        // Epic or Legendary: stage 1 — fog
-        await interaction.editReply({ embeds: [fogEmbed] });
-        await wait(1500);
-        // Stage 2 — partial reveal
-        const midColor = tierNum === 6 ? '#e74c3c' : tierNum === 4 ? '#9c27b0' : '#ff9800';
-        const midTitle = tierNum === 6 ? '☄️ Every animal in the forest has gone silent...' : tierNum === 4 ? '🔮 Something exceptional emerges...' : '⚡ The air crackles with power...';
-        const midTierLabel = tierNum === 6 ? 'EVENT' : tierNum === 4 ? 'EPIC' : 'LEGENDARY';
-        const midEmbed = new EmbedBuilder()
-            .setColor(midColor)
-            .setTitle(midTitle)
-            .setDescription(`━━━━━━━━━━━━━━━\n❓❓❓  **${midTierLabel}**  ❓❓❓\n━━━━━━━━━━━━━━━`);
-        await interaction.editReply({ embeds: [midEmbed] });
-        await wait(1500);
-        if (tierNum >= 5) {
-            // Stage 3 — legendary fanfare
-            const isEvent = tierNum === 6;
-            const fanfareEmbed = new EmbedBuilder()
-                .setColor(isEvent ? '#e74c3c' : '#ff9800')
-                .setTitle(isEvent ? '☄️ ⚡ 𝗠 𝗬 𝗧 𝗛 𝗜 𝗖 𝗔 𝗟 ⚡ ☄️' : '⚡ ✨ 𝗟 𝗘 𝗚 𝗘 𝗡 𝗗 𝗔 𝗥 𝗬 ✨ ⚡')
-                .setDescription(isEvent ? '━━━━━━━━━━━━━━━\n*Nothing like this has been seen in living memory.*\n━━━━━━━━━━━━━━━' : '━━━━━━━━━━━━━━━\n*The air crackles. This is once in a lifetime.*\n━━━━━━━━━━━━━━━');
-            await interaction.editReply({ embeds: [fanfareEmbed] });
-            await wait(1500);
-        }
-    }
-    await interaction.editReply({ embeds: [finalEmbed] });
-}
+const { stagedLootReveal } = require('../../../utils/stagedLootReveal');
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // START (was /hunt)
@@ -429,7 +384,7 @@ async function executeStart(interaction) {
         // Staged loot reveal for rare+ drops. A quick hunt skips the ceremony
         // here too — the fog-and-fanfare build-up is the same forced wait the
         // player opted out of, and the tier is still announced on the embed.
-        await stagedLootReveal(interaction, !quick && result.success ? result.tier : null, embed);
+        await stagedLootReveal(interaction, !quick && result.success ? result.tier : null, embed, 'hunt');
 
         if (result.success && ['epic', 'legendary', 'event'].includes(result.tier) && guildSettings?.economy?.announceRareDrops !== false) {
             const announceChannelId = guildSettings?.economy?.announcementChannelId;
@@ -733,5 +688,4 @@ function replyHuntPreflightFailure(interaction, preflight) {
 module.exports = {
     executeStart,
     replyHuntPreflightFailure,
-    stagedLootReveal,
 };
